requirements.txt
streamlit
opencv-python
numpy
imutils
easyocr
Pillow
matplotlib